Add: config/permissions.py 추가
여러 app에서 사용하기 위해 config 디렉토리에 permissions 파일 추가
기존 user/views.py에 있던 custom permission remove to config/permissions.py
