+++ user/views.py
@@ -7,30 +7,12 @@ from rest_framework import permissions
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
+from config.permissions import IsAdminOrNotAuthenticatedCreateOnly
 from user.models import User as UserModel
 from user.serializers import UserSerializer, UserDetailSerializer
 
+
 """"""
-# create custom permission
-class IsAdminOrNotAuthenticatedCreateOnly(permissions.BasePermission):
-    """
-    로그인 하지 않은 사용자만 post 가능
-    관리자는 모두 가능
-    """
-    SAFE_METHODS = ('POST', )
-    message = '접근 권한이 없습니다.'
-
-    def has_permission(self, request, view):
-        user = request.user
-
-        if request.method in self.SAFE_METHODS and not user.is_authenticated:
-            return True
-        elif user.is_admin:
-            return True
-
-        return False
-
-
 # Create your views here.
 class UserApiView(APIView):
     permission_classes = [IsAdminOrNotAuthenticatedCreateOnly]
